# docs/swagger.yaml
basePath: /api
definitions:
  core.AuthLogin:
    properties:
      phone:
        type: string
    required:
    - phone
    type: object
  core.AuthRegister:
    properties:
      phone:
        type: string
      username:
        minLength: 2
        type: string
    required:
    - phone
    - username
    type: object
...
  core.GetAllUserAvatarsResp:
    properties:
      avatar_id:
        type: integer
      avatar_url:
        type: string
    type: object
  core.GetProfileResp:
    properties:
      avatar_url:
        type: string
      name:
        type: string
      phone:
        type: string
      user_id:
        type: integer
    type: object
...
  core.UpdateUserReq:
    properties:
      name:
        minLength: 2
        type: string
      phone:
        type: string
    required:
    - name
    - phone
    type: object
...
  rest.errorResponse:
    properties:
      error:
        type: string
    type: object
  rest.tokenResponse:
    properties:
      token:
        type: string
    type: object
host: 35.180.225.96:8080
info:
  contact: {}
  description: API Server
  title: Social Network Backend
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: login
      operationId: login
      parameters:
      - description: credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/core.AuthLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      summary: Login
      tags:
      - Auth
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      description: refresh
      operationId: refresh
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      summary: Refresh
      tags:
      - Auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: register
      operationId: register
      parameters:
      - description: credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/core.AuthRegister'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      summary: Register
      tags:
      - Auth
  /api/auth/verify/{code}:
    get:
      consumes:
      - application/json
      description: verify
      operationId: verify
      parameters:
      - description: code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      summary: Verify
      tags:
      - Auth
...
  /api/profile/:
    get:
      description: get profile
      operationId: getProfile
      parameters:
      - description: profile
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/core.GetProfileResp'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: GetProfile
      tags:
      - Profile
  /api/profile/avatar/delete/{avatarId}:
    delete:
      description: delete avatar by avatar id
      operationId: deleteAvatar
      parameters:
      - description: avatarId
        in: path
        name: avatarId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: DeleteAvatar
      tags:
      - Profile
  /api/profile/avatar/upload:
    post:
      description: upload avatar
      operationId: uploadAvatar
      parameters:
      - description: avatar
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: UploadAvatar
      tags:
      - Profile
  /api/profile/avatars:
    get:
      description: get all avatars
      operationId: getAvatars
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/core.GetAllUserAvatarsResp'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: GetAvatars
      tags:
      - Profile
  /api/profile/update:
    patch:
      consumes:
      - application/json
      description: update profile
      operationId: updateProfile
      parameters:
      - description: user data to update
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/core.UpdateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: UpdateProfile
      tags:
      - Profile
  /api/stream/connect:
    get:
      description: connect to a streaming session
      operationId: streamConnect
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Connect
      tags:
      - Stream
  /api/stream/disconnect:
    post:
      description: disconnect from the streaming session
      operationId: streamDisconnect
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Disconnect
      tags:
      - Stream
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"